% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ig_components.R
\name{extract_ig_components}
\alias{extract_ig_components}
\title{Extract Institutional Grammar Components}
\usage{
extract_ig_components(tokens)
}
\arguments{
\item{tokens}{Vektor karakter yang berisi token teks yang akan dianalisis.}
}
\value{
Sebuah list dengan 17 elemen, masing-masing mewakili kategori IG 2.0.
}
\description{
Fungsi ini digunakan untuk mengekstrak komponen **Institutional Grammar (IG)** dari token teks
berdasarkan pola regex yang telah didefinisikan.
}
\examples{
tokens <- c("Gubernur", "harus", "menetapkan", "Rencana", "dalam waktu 6 bulan", "berdasarkan Undang-Undang")
result <- extract_ig_components(tokens)
print(result)

}
