% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_wordcloud.R
\name{generate_wordcloud}
\alias{generate_wordcloud}
\title{Word Cloud dengan Filter Frekuensi dan Stopwords}
\usage{
generate_wordcloud(text, min_freq = 1, stopwords = TRUE)
}
\arguments{
\item{text}{String teks yang ingin divisualisasikan dalam Word Cloud.}

\item{min_freq}{Frekuensi minimum kata untuk dimasukkan dalam Word Cloud.}

\item{stopwords}{Boolean, jika TRUE maka stopwords (kata umum) akan dihapus.}
}
\value{
Grafik Word Cloud menggunakan wordcloud2.
}
\description{
Fungsi ini membuat visualisasi Word Cloud dengan opsi untuk memfilter kata berdasarkan frekuensi dan mengaktifkan stopwords.
}
\examples{
text <- "Presiden menetapkan kebijakan nasional. Gubernur mengelola program daerah. Presiden menetapkan rencana."
generate_wordcloud(text, min_freq = 2, stopwords = TRUE)

}
