% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_stats.R
\name{net_stats}
\alias{net_stats}
\title{Statistik Jejaring untuk Komponen IG}
\usage{
net_stats(ig_data, component)
}
\arguments{
\item{ig_data}{List hasil dari fungsi `extract_ig_components`.}

\item{component}{Nama kategori IG yang ingin dihitung statistik jejaringnya.}
}
\value{
Statistik jejaring dalam bentuk list.
}
\description{
Fungsi ini menghitung statistik jejaring untuk kategori IG yang dipilih (misalnya, `Attribute`, `Deontic`, dll.).
}
\examples{
tokens <- c("Gubernur", "Menteri", "Presiden", "Bupati", "Direktur", "harus", "menetapkan")
ig_data <- extract_ig_components(tokens)
stats <- net_stats(ig_data, component = "Attribute")
print(stats)

}
