% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize_text.R
\name{tokenize_text}
\alias{tokenize_text}
\title{Tokenisasi Teks}
\usage{
tokenize_text(text)
}
\arguments{
\item{text}{String yang berisi teks yang akan ditokenisasi.}
}
\value{
Vektor karakter yang berisi token yang dihasilkan dari teks input.
}
\description{
Fungsi ini digunakan untuk melakukan tokenisasi teks berbahasa Indonesia. Proses tokenisasi mencakup
normalisasi huruf kecil, penghapusan karakter tidak relevan, dan pemisahan teks menjadi token berdasarkan spasi.
}
\examples{
text <- "Gubernur menetapkan kebijakan sesuai dengan aturan."
tokens <- tokenize_text(text)
print(tokens)

}
