% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_attribute.R
\name{visualize_attribute}
\alias{visualize_attribute}
\title{Visualisasi Komponen Attribute}
\usage{
visualize_attribute(ig_data)
}
\arguments{
\item{ig_data}{List hasil dari fungsi `extract_ig_components`.}
}
\value{
Plot jaringan `Attribute`.
}
\description{
Fungsi ini digunakan untuk memvisualisasikan token dalam komponen `Attribute` menggunakan pustaka `igraph`.
}
\examples{
tokens <- c("Gubernur", "harus", "menetapkan", "Rencana", "Menteri", "Presiden")
ig_data <- extract_ig_components(tokens)
visualize_attribute(ig_data)

}
