% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_attribute_aim.R
\name{visualize_attribute_aim}
\alias{visualize_attribute_aim}
\title{Jejaring antara Attribute dan Aim}
\usage{
visualize_attribute_aim(ig_data)
}
\arguments{
\item{ig_data}{List hasil dari fungsi `extract_ig_components`.}
}
\value{
Grafik jejaring antara Attribute dan Aim.
}
\description{
Fungsi ini membuat jejaring antara komponen `Attribute` dan `Aim` menggunakan pustaka `igraph`.
}
\examples{
tokens <- c("Gubernur", "Menteri", "menetapkan", "mengelola", "Direktur")
ig_data <- extract_ig_components(tokens)
visualize_attribute_aim(ig_data)

}
