% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_filtered_attribute_aim.R
\name{visualize_filtered_attribute_aim}
\alias{visualize_filtered_attribute_aim}
\title{Visualisasi Jejaring Attribute dan Aim dengan Filter}
\usage{
visualize_filtered_attribute_aim(ig_data, attribute = NULL, aim = NULL)
}
\arguments{
\item{ig_data}{List hasil dari fungsi `extract_ig_components`.}

\item{attribute}{Nama `Attribute` yang ingin divisualisasikan (opsional).}

\item{aim}{Nama `Aim` yang ingin divisualisasikan (opsional).}
}
\value{
Grafik jejaring untuk kombinasi Attribute dan Aim yang dipilih.
}
\description{
Fungsi ini membuat jejaring hanya untuk kombinasi `Attribute` dan `Aim` tertentu.
}
\examples{
tokens <- c("Presiden", "Gubernur", "menetapkan", "melaksanakan", "Rencana")
ig_data <- extract_ig_components(tokens)
visualize_filtered_attribute_aim(ig_data, attribute = "Presiden", aim = "melaksanakan")

}
