% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_ig.R
\name{visualize_ig}
\alias{visualize_ig}
\title{Visualisasi Institutional Grammar}
\usage{
visualize_ig(ig_data)
}
\arguments{
\item{ig_data}{List hasil dari fungsi `extract_ig_components`, yang berisi kategori IG sebagai kunci
dan token-token terkait sebagai elemen.}
}
\value{
Plot jaringan IG.
}
\description{
Fungsi ini digunakan untuk memvisualisasikan data IG menggunakan pustaka `igraph`.
}
\examples{
tokens <- c("Gubernur", "harus", "menetapkan", "Rencana", "dalam waktu 6 bulan", "berdasarkan Undang-Undang")
ig_data <- extract_ig_components(tokens)
visualize_ig(ig_data)

}
