% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_specific_aim.R
\name{visualize_specific_aim}
\alias{visualize_specific_aim}
\title{Visualisasi Aim Tertentu}
\usage{
visualize_specific_aim(ig_data, aim)
}
\arguments{
\item{ig_data}{List hasil dari fungsi `extract_ig_components`.}

\item{aim}{Nama `Aim` yang ingin divisualisasikan (contoh: "menetapkan").}
}
\value{
Grafik jejaring untuk `Aim` yang dipilih.
}
\description{
Fungsi ini membuat jejaring hanya untuk token yang memiliki `Aim` tertentu.
}
\examples{
tokens <- c("Gubernur", "Menteri", "menetapkan", "mengelola", "Rencana")
ig_data <- extract_ig_components(tokens)
visualize_specific_aim(ig_data, aim = "menetapkan")

}
